# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music Distribution CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List all artists
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: label_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedArtists"
    post:
      operationId: createArtist
      tags: [artists]
      summary: Create a new artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistBody"
      responses:
        "201":
          description: Artist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get artist by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDetail"
    put:
      operationId: updateArtist
      tags: [artists]
      summary: Update artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistBody"
      responses:
        "200":
          description: Artist updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
    delete:
      operationId: deleteArtist
      tags: [artists]
      summary: Delete artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Artist deleted
  /artists/{id}/stats:
    get:
      operationId: getArtistStats
      tags: [artists]
      summary: Get artist analytics stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistStats"

  # ==================== LABELS ====================
...
  /crm/contacts:
    get:
      operationId: listContacts
      tags: [crm-contacts]
      summary: List CRM contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [artist, author, label, manager, partner]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedContacts"
    post:
      operationId: createContact
      tags: [crm-contacts]
      summary: Create CRM contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /crm/contacts/{id}:
    get:
      operationId: getContact
      tags: [crm-contacts]
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [crm-contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [crm-contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contact deleted

  # ==================== CRM TASKS ====================
  /crm/tasks:
    get:
      operationId: listTasks
      tags: [crm-tasks]
      summary: List CRM tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: assigned_to
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCrmTasks"
    post:
      operationId: createCrmTask
      tags: [crm-tasks]
      summary: Create CRM task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrmTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
  /crm/tasks/{id}:
    get:
      operationId: getCrmTask
      tags: [crm-tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
    put:
      operationId: updateCrmTask
      tags: [crm-tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrmTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
    delete:
      operationId: deleteCrmTask
      tags: [crm-tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

  # ==================== FINANCE / TRANSACTIONS ====================
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ==================== DASHBOARD ====================
...
    Artist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        labelId:
          type: integer
          nullable: true
        labelName:
          type: string
          nullable: true
        spotifyId:
          type: string
          nullable: true
        appleId:
          type: string
          nullable: true
        socialLinks:
          type: object
          nullable: true
        status:
          type: string
        totalReleases:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, slug, status, totalReleases, createdAt, updatedAt]
    ArtistDetail:
      allOf:
        - $ref: "#/components/schemas/Artist"
        - type: object
          properties:
            releases:
              type: array
              items:
                $ref: "#/components/schemas/Release"
            recentTracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"
    ArtistStats:
      type: object
      properties:
        totalStreams:
          type: integer
        totalRevenue:
          type: number
        monthlyListeners:
          type: integer
        topPlatform:
          type: string
        streamsByPlatform:
          type: array
          items:
            $ref: "#/components/schemas/PlatformAnalytics"
      required: [totalStreams, totalRevenue, monthlyListeners, topPlatform, streamsByPlatform]
    CreateArtistBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        labelId:
          type: integer
          nullable: true
        spotifyId:
          type: string
          nullable: true
        appleId:
          type: string
          nullable: true
        socialLinks:
          type: object
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          default: active
      required: [name]
    PaginatedArtists:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Artist"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # ==================== LABELS ====================
...
    Release:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        releaseType:
          type: string
          enum: [single, album, ep, compilation]
        status:
          type: string
          enum: [draft, pending_review, approved, delivering, delivered, live, error, takedown_requested, removed]
        upc:
          type: string
          nullable: true
        artistId:
          type: integer
        artistName:
          type: string
        labelId:
          type: integer
          nullable: true
        labelName:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        releaseDate:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        isExplicit:
          type: boolean
        territories:
          type: array
          items:
            type: string
        totalTracks:
          type: integer
        pLine:
          type: string
          nullable: true
        cLine:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, releaseType, status, artistId, artistName, isExplicit, territories, totalTracks, createdAt, updatedAt]
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        isrc:
          type: string
          nullable: true
        releaseId:
          type: integer
          nullable: true
        releaseName:
          type: string
          nullable: true
        artistId:
          type: integer
        artistName:
          type: string
        trackNumber:
          type: integer
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        genre:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        isExplicit:
          type: boolean
        composerName:
          type: string
          nullable: true
        lyricistName:
          type: string
          nullable: true
        iswc:
          type: string
          nullable: true
        audioUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, artistId, artistName, isExplicit, createdAt, updatedAt]
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [artist, author, label, manager, partner]
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        telegram:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, createdAt, updatedAt]
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [artist, author, label, manager, partner]
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        telegram:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [name, type]
    PaginatedContacts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # ==================== CRM TASKS ====================
    CrmTask:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        relatedEntityType:
          type: string
          nullable: true
        relatedEntityId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, priority, createdAt, updatedAt]
    CreateCrmTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
          default: todo
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        assignedToId:
          type: integer
          nullable: true
        dueDate:
          type: string
          nullable: true
        relatedEntityType:
          type: string
          nullable: true
        relatedEntityId:
          type: integer
          nullable: true
      required: [title]
    PaginatedCrmTasks:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CrmTask"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # ==================== FINANCE ====================
...
    PlatformAnalytics:
      type: object
      properties:
        platform:
          type: string
        streams:
          type: integer
        revenue:
          type: number
        percentage:
          type: number
      required: [platform, streams, revenue, percentage]
...