# lib/api-zod/src/generated/types/release.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Distribution CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ReleaseReleaseType } from "./releaseReleaseType";
import type { ReleaseStatus } from "./releaseStatus";

export interface Release {
  id: number;
  title: string;
  releaseType: ReleaseReleaseType;
  status: ReleaseStatus;
  upc?: string | null;
  artistId: number;
  artistName: string;
  labelId?: number | null;
  labelName?: string | null;
  coverUrl?: string | null;
  genre?: string | null;
  releaseDate?: string | null;
  language?: string | null;
  isExplicit: boolean;
  territories: string[];
  totalTracks: number;
  pLine?: string | null;
  cLine?: string | null;
  createdAt: string;
  updatedAt: string;
}